style : scanbasePackage를 최상위 루트에서 읽을 수 있도록 패키지 구조를 변경
