+++ new_goatodo/goatodo/application/src/test/java/com/goatodo/ApplicationTests.java
@@ -1,4 +1,4 @@
-package com.goatodo.application;
+package com.goatodo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -7,8 +7,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EntityScan(basePackages = {"com.goatodo.domain"})
-@EnableJpaRepositories(basePackages = {"com.goatodo.domain"})
 class ApplicationTests {
 
     @Test
