+++ new_goatodo/goatodo/domain/src/test/java/com/goatodo/DomainApplicationTests.java
@@ -1,10 +1,11 @@
-package com.goatodo.web;
+package com.goatodo;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class WebApplicationTests {
+@SpringBootApplication
+class DomainApplicationTests {
 
     @Test
     void contextLoads() {
