--- new_goatodo/goatodo/domain/src/test/java/com/goatodo/domain/DomainApplicationTests.java
@@ -1,14 +1,0 @@
-package com.goatodo.domain;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootApplication
-class DomainApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
