+++ new_goatodo/goatodo/web/src/main/java/com/goatodo/GoatodoApplication.java
@@ -1,0 +1,13 @@
+package com.goatodo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GoatodoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GoatodoApplication.class, args);
+    }
+
+}
