--- new_goatodo/goatodo/web/src/main/java/com/goatodo/web/WebApplication.java
@@ -1,14 +1,0 @@
-package com.goatodo.web;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication(scanBasePackages = {"com.goatodo"})
-public class WebApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(WebApplication.class, args);
-    }
-
-}
