+++ new_goatodo/goatodo/web/src/test/java/com/goatodo/GoatodoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.goatodo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GoatodoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
